Fix root-window-bkgn-theme-mismatch bug
When creating Application, the background of the root window (i.e.
the scene color of the renderer) was not synchronized with the
initialized theme. For example, the background would keep white
even initialized with dark mode. Moreover, the background would
not synchronize with the theme mode either if we created the
application without a main window. The reason is very simple:
the code to synchronize the background with the theme was placed
into onChangeTheme callback of MainWindow. To fix the bug, we
move the code into Constructor and ChangeTheme of Application.